requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cffi==1.16.0
charset-normalizer==3.3.0
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.26.1
pycparser==2.21
soundfile==0.12.1
yarl==1.9.2
mido==1.3.0
surgepy @ file:///path/to/surge/src/surge-python
